requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
